+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Run integration tests in a separate profile
 * Updated Jackson JSON library to fix null-Enum bug
 * Added support for "next" numbers in settings
+* Added support for BCC addresses in settings
 
 ## 1.0.2 (2012-12-21)